+++ module-graph/src/main/java/ch/hsr/adv/ui/graph/logic/GraphParser.java
@@ -7,12 +7,12 @@ import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVValueStyle;
 import ch.hsr.adv.commons.graph.logic.ConstantsGraph;
+import ch.hsr.adv.commons.graph.logic.domain.GraphElement;
+import ch.hsr.adv.commons.graph.logic.domain.GraphRelation;
 import ch.hsr.adv.ui.core.logic.GsonProvider;
 import ch.hsr.adv.ui.core.logic.InterfaceAdapter;
 import ch.hsr.adv.ui.core.logic.Parser;
 import ch.hsr.adv.ui.core.logic.exceptions.ADVParseException;
-import ch.hsr.adv.ui.graph.logic.domain.GraphElement;
-import ch.hsr.adv.ui.graph.logic.domain.GraphRelation;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
